misc/HtmlTestSuite/results/test-0658.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://bugs.chromium.org/p/chromium/issues/detail?id=1127415" rel="help"/>
<meta content="The presence of a scrollbar should not make a min-sized flexbox, and a single flex-item differ in size." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div style="display: flex; width: 100px;">
  <div id="outer" style="display: flex; width: min-content; height: 100px;">
    <div id="inner" style="flex: 0 0 auto; overflow-y: auto;">
      <div style="width: 100px; height: 3000px;"/>
    </div>
  </div>
</div>
<script>
test(() =&gt; {
  const outer = document.getElementById('outer');
  const inner = document.getElementById('inner');
  assert_equals(outer.getBoundingClientRect().width, inner.getBoundingClientRect().width);
});
</script>
</body></html>
